Add user fk

+++ app/Http/Controllers/EntryController.php
@@ -4,9 +4,11 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\EntryRequest;
 use App\Models\Entry;
+use App\Models\User;
 use Exception;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
 use Illuminate\Validation\ValidationException;
 use RuntimeException;
 use Symfony\Component\HttpFoundation\Response;
@@ -14,11 +16,14 @@ use Symfony\Component\HttpFoundation\Response;
 class EntryController extends Controller
 {
     /**
+     * @param Request $request
      * @return JsonResponse
      */
-    public function index(): JsonResponse
+    public function index(Request $request): JsonResponse
     {
-        $entries = Entry::all();
+        /** @var User $user */
+        $user = $request->user();
+        $entries = Entry::where('user_id', $user->id)->get();
 
         return new JsonResponse($entries, Response::HTTP_OK);
     }
@@ -34,7 +39,10 @@ class EntryController extends Controller
         } /** @noinspection PhpRedundantCatchClauseInspection */ catch (ValidationException $e) {
             return new JsonResponse($e, Response::HTTP_UNPROCESSABLE_ENTITY);
         }
+        /** @var User $user */
+        $user = $request->user();
         $entry = new Entry();
+        $entry->userId = $user->id;
         $entry->label = $validated['label'];
         $entry->value = $validated['value'];
         if ($validated['date']) {
@@ -63,6 +71,11 @@ class EntryController extends Controller
         } catch (ModelNotFoundException $e) {
             return new JsonResponse($e, Response::HTTP_NOT_FOUND);
         }
+        /** @var User $user */
+        $user = $request->user();
+        if ($user->id != $entry->userId) {
+            return new JsonResponse(null, Response::HTTP_NOT_FOUND);
+        }
         $entry->label = $validated['label'];
         $entry->value = $validated['value'];
         if ($validated['date']) {
@@ -75,15 +88,21 @@ class EntryController extends Controller
 
     /**
      * @param int $id
+     * @param Request $request
      * @return JsonResponse
      */
-    public function destroy(int $id): JsonResponse
+    public function destroy(int $id, Request $request): JsonResponse
     {
         try {
             /** @var Entry $entry */
             $entry = Entry::find($id);
         } catch (ModelNotFoundException $e) {
             return new JsonResponse($e, Response::HTTP_NOT_FOUND);
+        }
+        /** @var User $user */
+        $user = $request->user();
+        if ($user->id != $entry->userId) {
+            return new JsonResponse(null, Response::HTTP_NOT_FOUND);
         }
         try {
             $entry->delete();
